🏗️ Adjusting package name for krpc-code-gen
